LayoutTests/inspector/debugger/set-breakpoint-expected.txt

Tests setting breakpoints.

Debugger was enabled.

Running: testSetBreakpoint
Script source was shown.
Set timer for test function.
Script execution paused.
Call stack:
    0) testFunction (set-breakpoint.html:17)

Dump breakpoint sidebar pane:

    set-breakpoint.html:17
    var x = Math.sqrt(10);

Dump breakpoint sidebar pane:

Script execution resumed.

Running: testSetBreakpointOnTheLastLine
Call stack:
    0) oneLineTestFunction (set-breakpoint.html:7)

Running: testSetBreakpointOnTheLastLine2
Call stack:
    0) oneLineTestFunction2 (set-breakpoint.html:11)

Running: testSetBreakpointOnTheSameLine
Debugger was disabled.
